NATIVE-532: Move webviewFontFamilies.js
